+++ src/referencedata-period/period-data-builder.spec.js
@@ -1,0 +1,67 @@
+/*
+ * This program is part of the OpenLMIS logistics management information system platform software.
+ * Copyright © 2017 VillageReach
+ *
+ * This program is free software: you can redistribute it and/or modify it under the terms
+ * of the GNU Affero General Public License as published by the Free Software Foundation, either
+ * version 3 of the License, or (at your option) any later version.
+ *  
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
+ * See the GNU Affero General Public License for more details. You should have received a copy of
+ * the GNU Affero General Public License along with this program. If not, see
+ * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org. 
+ */
+
+(function() {
+
+    'use strict';
+
+
+    angular
+        .module('referencedata-period')
+        .factory('PeriodDataBuilder', PeriodDataBuilder);
+
+    PeriodDataBuilder.$inject = ['Period', 'ScheduleDataBuilder'];
+
+    function PeriodDataBuilder(Period, ScheduleDataBuilder) {
+
+        PeriodDataBuilder.prototype.withStartDate = withStartDate;
+        PeriodDataBuilder.prototype.withEndDate = withEndDate;
+        PeriodDataBuilder.prototype.build = build;
+
+        return PeriodDataBuilder;
+
+        function PeriodDataBuilder() {
+            this.id = '516ac930-0d28-49f5-a178-64764e22b236';
+            this.name = 'Jan2017';
+            this.description = 'description';
+            this.startDate = new Date(2017, 0, 1);
+            this.endDate = new Date(2017, 0, 31);
+            this.processingSchedule = new ScheduleDataBuilder().build();
+        }
+
+        function withStartDate(newDate) {
+            this.startDate = newDate;
+            return this;
+        }
+
+        function withEndDate(newDate) {
+            this.endDate = newDate;
+            return this;
+        }
+
+        function build() {
+            return new Period(
+                this.id,
+                this.name,
+                this.description,
+                this.startDate,
+                this.endDate,
+                this.processingSchedule
+            );
+        }
+
+    }
+
+})();